+++ tests/unit/dependency_getter/test_pdm.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from deptry.dependency_getter.pdm import PDMDependencyGetter
+from deptry.dependency_getter.pep621.pdm import PDMDependencyGetter
 from tests.utils import run_within_dir
